automail-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
transformers>=4.21.0
torch>=1.12.0
python-dotenv==1.0.0
requests==2.31.0
sentencepiece>=0.1.97 